remove pooling

+++ src/napatrackmater/Trackvector.py
@@ -2066,7 +2066,7 @@ def predict_with_model(saved_model_path, saved_model_json, features_array, thres
     for idx in range(features_array.shape[0]):
         feature_type = features_array[idx,:].tolist()
         feature_type = z_score_normalization(feature_type)
-        features_tensor = torch.tensor(feature_type, dtype=torch.float32).to(device)
+        features_tensor = torch.tensor(feature_type).to(device)
         new_data_with_channel = features_tensor.unsqueeze(0).unsqueeze(0)
        
         with torch.no_grad():
